% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinGRanges.r
\name{BinGRanges}
\alias{BinGRanges}
\title{Bin a GRanges.}
\usage{
BinGRanges(
  gRange.gnr = NULL,
  chromSize.dtf = NULL,
  binSize.num = NULL,
  method.chr = "mean",
  variablesName.chr_vec = NULL,
  na.rm = TRUE,
  cores.num = 1,
  reduce.bln = TRUE,
  verbose.bln = TRUE
)
}
\arguments{
\item{gRange.gnr}{<GRanges>: a GRanges to bin.}

\item{chromSize.dtf}{<data.frame>: a data.frame where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.}

\item{binSize.num}{<numerical>: width of the bins.}

\item{method.chr}{<character>: name of a summary method as 'mean', 'median', 'sum', 'max, 'min'. (Default 'mean')}

\item{variablesName.chr_vec}{<character> : a character vector that specify the metadata columns of GRanges on which apply the summary method.}

\item{na.rm}{<logical> : a logical value indicating whether 'NA' values should be stripped before the computation proceeds. (Default TRUE)}

\item{cores.num}{<numerical> : the number of cores. (Default 1)}

\item{reduce.bln}{<logical> : whether duplicated Bin must been reduced with de summary method. (Default TRUE)}

\item{verbose.bln}{<logical>: if TRUE show the progression in console. (Default TRUE)}
}
\value{
A binned GRanges.
}
\description{
BinGRanges
Bin a GRanges and allow to apply a summary method (e.g: 'mean', 'median', 'sum', 'max, 'min' ...) to a chossen numericals variables of ranges in a same bin.
}
\examples{
GRange.gnr <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr1", "chr2"), c(3, 1)),
    ranges = IRanges::IRanges(c(1,201,251,1), end = c(200,250,330,100), names = letters[1:4]),
    strand = S4Vectors::Rle(BiocGenerics::strand(c("*")), 4),
    score = c(50,NA,100,30)
    )
GRange.gnr
chromSize.dtf = data.frame(c("chr1","chr2"),c(350,100))
binSize.num <- 100
binnedGRanges.gnr <- BinGRanges(
    gRange.gnr = GRange.gnr,
    chromSize.dtf=chromSize.dtf,
    binSize.num=binSize.num,
    method.chr ="mean",
    variablesName.chr_vec="score",
    na.rm=TRUE
)
binnedGRanges.gnr
}
