% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicSystem.r
\name{GenomicSystem}
\alias{GenomicSystem}
\title{Convert numbers of base pairs into string and vice versa.}
\usage{
GenomicSystem(x, digits.num = 3)
}
\arguments{
\item{x}{<character or numeric>: the number to convert or string to convert.}

\item{digits.num}{<integer>: The number of significant digits to be used. See signif() for more informations. (Default 3)}
}
\value{
the converted number or string.
}
\description{
GenomicSystem
Convert numbers of base into string with order of magnitude (Kbp, Mbp, Gbp) and vice versa.
}
\examples{
GenomicSystem(1540,3)
GenomicSystem(1540,2)
GenomicSystem("1Mbp")
GenomicSystem("1Kbp")
GenomicSystem("1k")
}
