% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexGRanges.r
\name{IndexGRanges}
\alias{IndexGRanges}
\title{IndexGRanges}
\usage{
IndexGRanges(
  gRange.gnr_lst = NULL,
  constraint.gnr = NULL,
  chromSize.dtf = NULL,
  binSize.int = NULL,
  method.chr = "mean",
  variablesName.chr_vec = NULL,
  cores.num = 1,
  verbose.bln = TRUE
)
}
\arguments{
\item{gRange.gnr_lst}{<GRanges or GRangesList or list[GRanges]>: some GRanges or a list of GRanges or a GRangesList}

\item{constraint.gnr}{<GRanges>: A GRange of constraint regions. If NULL chromosomes in chromSize.dtf are used (Default NULL)}

\item{chromSize.dtf}{<data.frame>: A data.frame of genome where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.}

\item{binSize.int}{<integer>: A number that specify the width bins.}

\item{method.chr}{<character>: A string of a summary method name as 'mean', 'median', 'sum', 'max, 'min'. (Default 'mean'')}

\item{variablesName.chr_vec}{<character> : A character vector that specify the metadata columns of GRanges on which apply the summary method.}

\item{cores.num}{<integer> : An integer to specify the number of cores. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
A GRange object.
}
\description{
Bin multiple GRanges and summary all in one GRange. Could overlap ranges with constraints regions
}
